Fixed absent exception handling, controller related mistakes and reformatted code

+++ article-analyzer/src/main/java/co/inventorsoft/academy/spring/services/UserService.java
@@ -1,6 +1,7 @@
 package co.inventorsoft.academy.spring.services;
 
-import co.inventorsoft.academy.spring.models.NotificationType;
+import co.inventorsoft.academy.spring.exceptions.NotFoundException;
+import co.inventorsoft.academy.spring.exceptions.ValidationException;
 import co.inventorsoft.academy.spring.models.User;
 import co.inventorsoft.academy.spring.repositories.UserJsonRepository;
 import co.inventorsoft.academy.spring.repositories.UserRepository;
@@ -35,25 +36,42 @@ public class UserService {
     }
 
     public User getUserById(Long userId){
-        return userRepository.findById(userId).orElseGet(() -> User.builder()
-            .email("userNotFound@mail.com")
-            .build()
-        );
+        Optional<User> currentUser = userRepository.findById(userId);
+
+        if(currentUser.isEmpty()){
+            throw new NotFoundException(NotFoundException.USER_ID_NOT_FOUND);
+        }
+
+        return currentUser.get();
     }
 
     public User getUserByEmail(String userEmail){
-        return userRepository.findByEmail(userEmail).orElseGet(() -> User.builder()
-            .email("userNotFound@mail.com")
-            .build()
-        );
+        Optional<User> currentUser = userRepository.findByEmail(userEmail);
+
+        if(currentUser.isEmpty()){
+            throw new NotFoundException(NotFoundException.USER_EMAIL_NOT_FOUND);
+        }
+
+        return currentUser.get();
     }
 
     public User createUser(User user){
+        if(userRepository.findById(user.getId()).isPresent()){
+            throw new ValidationException(ValidationException.USER_ID_EXISTS);
+        }
+
+        if(checkUser(user)){
+            throw new ValidationException(ValidationException.USER_MUST_NOT_HAVE_NULL_FIELDS);
+        }
 
         return userRepository.save(user);
     }
 
     public User updateUser(Long id, User newUser){
+        if(checkUser(newUser)){
+            throw new ValidationException(ValidationException.USER_MUST_NOT_HAVE_NULL_FIELDS);
+        }
+
         User updatedUser = getUserById(id);
         updatedUser.setEmail(newUser.getEmail());
         updatedUser.setSlackId(newUser.getSlackId());
@@ -69,10 +87,18 @@ public class UserService {
     }
 
     public void deleteUserById(Long userId){
+        if(userRepository.findById(userId).isEmpty()){
+            throw new ValidationException(ValidationException.USER_ID_DOES_NOT_EXIST);
+        }
+
         userRepository.deleteById(userId);
     }
 
-
+    private boolean checkUser(User userForCheck){
+        return userForCheck.getId() == null || userForCheck.getUsername() == null ||
+            userForCheck.getEmail() == null
+            || userForCheck.getSlackId() == null || userForCheck.getNotificationType() == null;
+    }
 
 
 }
